+++ java-generator/core/src/main/java/io/fabric8/java/generator/nodes/JObjectExtraAnnotations.java
@@ -24,7 +24,13 @@ public interface JObjectExtraAnnotations {
 
   default void addExtraAnnotations(ClassOrInterfaceDeclaration clz) {
     clz.addAnnotation("lombok.ToString");
-    clz.addAnnotation("lombok.EqualsAndHashCode");
+    if (clz.getExtendedTypes().isEmpty()) {
+      clz.addAnnotation("lombok.EqualsAndHashCode");
+    } else {
+      clz.addAnnotation(new SingleMemberAnnotationExpr(
+          new Name("lombok.EqualsAndHashCode"),
+          new NameExpr("callSuper = true")));
+    }
     clz.addAnnotation("lombok.Setter");
 
     clz.addAnnotation(