+++ kubernetes-client-api/src/test/java/io/fabric8/kubernetes/client/utils/SerializationTest.java
@@ -15,15 +15,21 @@
  */
 package io.fabric8.kubernetes.client.utils;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import com.fasterxml.jackson.annotation.JsonUnwrapped;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonTypeResolver;
 import com.fasterxml.jackson.databind.node.BooleanNode;
 import com.fasterxml.jackson.databind.node.IntNode;
 import com.fasterxml.jackson.databind.node.TextNode;
+import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.GenericKubernetesResource;
 import io.fabric8.kubernetes.api.model.KubernetesList;
 import io.fabric8.kubernetes.api.model.KubernetesResource;
 import io.fabric8.kubernetes.api.model.Namespace;
+import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.PodBuilder;
 import io.fabric8.kubernetes.api.model.PodSpec;
@@ -32,6 +38,8 @@ import io.fabric8.kubernetes.api.model.Service;
 import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.CustomResourceDefinition;
 import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.JSONSchemaProps;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
+import io.fabric8.kubernetes.api.model.coordination.v1.Lease;
+import io.fabric8.kubernetes.api.model.coordination.v1.LeaseSpec;
 import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.kubernetes.model.annotation.Group;
 import io.fabric8.kubernetes.model.annotation.Version;
@@ -45,11 +53,13 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
+import java.time.ZonedDateTime;
 import java.util.Collections;
 import java.util.Map;
 import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class SerializationTest {
@@ -320,6 +330,101 @@ class SerializationTest {
   void quantityQuoting() {
     Quantity quantity = Serialization.unmarshal("amount: \"2\"\nformat: \"Gi\"", Quantity.class);
     assertThat(Serialization.asYaml(quantity)).isEqualTo("--- \"2Gi\"\n");
+  }
+
+  @JsonTypeResolver(io.fabric8.kubernetes.model.jackson.UnwrappedTypeResolverBuilder.class)
+  @JsonInclude(JsonInclude.Include.NON_NULL)
+  @JsonSubTypes({
+      @JsonSubTypes.Type(Named.class),
+      @JsonSubTypes.Type(Counted.class)
+  })
+  @JsonTypeInfo(use = JsonTypeInfo.Id.DEDUCTION)
+  interface Poly {
+
+  }
+
+  @JsonDeserialize(using = com.fasterxml.jackson.databind.JsonDeserializer.None.class)
+  @JsonInclude(JsonInclude.Include.NON_NULL)
+  static class Named implements Poly {
+    public String name;
+  }
+
+  @JsonDeserialize(using = com.fasterxml.jackson.databind.JsonDeserializer.None.class)
+  static class Counted implements Poly {
+    public int count;
+  }
+
+  @JsonDeserialize(using = io.fabric8.kubernetes.model.jackson.JsonUnwrappedDeserializer.class)
+  @JsonInclude(JsonInclude.Include.NON_NULL)
+  static class PolyParent {
+    @JsonUnwrapped
+    public Poly poly;
+    public int other;
+  }
+
+  @Test
+  void polymorphicUnwrap() {
+    PolyParent parent = Serialization.unmarshal("other: 1\nname: x", PolyParent.class);
+    assertEquals("x", ((Named) parent.poly).name);
+
+    // should be null if it can't be deduced
+    parent = Serialization.unmarshal("other: 1", PolyParent.class);
+    assertThat(parent.poly).isNull();
+  }
+
+  @Test
+  void unmarshal_podWithYYYYMMDDAnnotation_shouldNotConvertToDate() {
+    // Given
+    InputStream inputStream = getClass().getResourceAsStream("/serialization/test-pod-manifest-dateformat-annotation.yml");
+
+    // When
+    Pod pod = Serialization.unmarshal(inputStream);
+
+    // Then
+    assertThat(pod)
+        .isNotNull()
+        .hasFieldOrPropertyWithValue("spec.shareProcessNamespace", false)
+        .hasFieldOrPropertyWithValue("spec.terminationGracePeriodSeconds", 2L)
+        .extracting(Pod::getMetadata)
+        .extracting(ObjectMeta::getAnnotations)
+        .isEqualTo(Collections.singletonMap("report_date", "2020-01-01"));
+  }
+
+  @Test
+  void unmarshal_leaseWithTypedValues_ShouldConvertToTypedFields() {
+    // Given
+    InputStream inputStream = getClass().getResourceAsStream("/serialization/test-lease-manifest-typed-fields.yml");
+
+    // When
+    Lease lease = Serialization.unmarshal(inputStream);
+
+    // Then
+    assertThat(lease)
+        .isNotNull()
+        .hasFieldOrPropertyWithValue("metadata.annotations",
+            Collections.singletonMap("creationTimestamp", "2022-07-22T10:42:02Z"))
+        .hasFieldOrPropertyWithValue("spec.leaseDurationSeconds", 2)
+        .hasFieldOrPropertyWithValue("spec.leaseTransitions", 3)
+        .extracting(Lease::getSpec)
+        .extracting(LeaseSpec::getAcquireTime)
+        .isInstanceOf(ZonedDateTime.class);
+  }
+
+  @Test
+  void unmarshal_configMapWithBoolAndDateData_shouldBeDeserializedAsString() {
+    // Given
+    InputStream inputStream = getClass().getResourceAsStream("/serialization/test-configmap-manifest-bool-dateformat.yml");
+
+    // When
+    ConfigMap configMap = Serialization.unmarshal(inputStream);
+
+    // Then
+    assertThat(configMap)
+        .isNotNull()
+        .extracting(ConfigMap::getData)
+        .asInstanceOf(InstanceOfAssertFactories.MAP)
+        .containsEntry("bool", "NO")
+        .containsEntry("date", "2022-07-22");
   }
 
 }