+++ CHANGELOG.md
@@ -3,17 +3,33 @@
 ### 6.1-SNAPSHOT
 
 #### Bugs
+* Fix #4256: crd-generator-apt pom.xml includes transitive dependencies
+* Fix #4294: crd-generator respects JsonIgnore annotations on enum properties
+* Fix #4247: NO_PROXY with invalid entries throws exception
 
 #### Improvements
+* Fix #4254: adding debug logging for exec stream messages
 * Fix #4041: adding Quantity.getNumericalAmount with an explanation about bytes and cores.
 * Fix #4241: added more context to informer logs with the endpoint path
 * Fix #887: added KubernetesClient.visitResources to search and perform other operations across all resources.
+* Fix #4250: allowing for deserialization of polymorphic unwrapped fields
+* Fix #4259: Java Generator's CR should have Lombok's `@EqualsAndHashCode` with `callSuper = true`
 
 #### Dependency Upgrade
 
 #### New Features
+* Fix #2271: Support periodic refresh of access tokens before they expire
 
 #### _**Note**_: Breaking changes in the API
+
+### 5.12.3 (2022-07-27)
+
+#### Bugs
+* Fix #3969: relist will not trigger sync events
+* Fix #4049: properly populate exception metadata with resource information if available
+* Fix #4222: backport of #4082 - to not process events until the cache is complete
+* Fix #4246: KubernetesClientException is swallowed in LeaderElector
+* Fix #4295: Configure SnakeYaml to ignore converting timestamps to Date objects
 
 ### 6.0.0 (2022-07-13)